+++ app/Http/Controllers/Admin/RolesController.php
@@ -5,10 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Permission;
 use App\Role;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\MassDestroyRoleRequest;
-use App\Http\Requests\StoreRoleRequest;
-use App\Http\Requests\UpdateRoleRequest;
 use App\Util\BasicFunctions;
+use Illuminate\Http\Request;
 
 class RolesController extends Controller
 {
@@ -32,10 +30,15 @@ class RolesController extends Controller
         return view('admin.shared.form_edit', compact('formEntries', 'route', 'header', 'method'));
     }
 
-    public function store(StoreRoleRequest $request)
+    public function store(Request $request)
     {
         abort_unless(\Gate::allows('role_create'), 403);
 
+        $request->validate([
+            'title' => 'required',
+            'permissions' => 'array',
+            'permissions.*' => 'integer',
+        ]);
         $role = Role::create($request->all());
         $role->permissions()->sync($request->input('permissions', []));
 
@@ -53,10 +56,15 @@ class RolesController extends Controller
         return view('admin.shared.form_edit', compact('formEntries', 'route', 'header', 'method'));
     }
 
-    public function update(UpdateRoleRequest $request, Role $role)
+    public function update(Request $request, Role $role)
     {
         abort_unless(\Gate::allows('role_edit'), 403);
 
+        $request->validate([
+            'title' => 'required',
+            'permissions' => 'array',
+            'permissions.*' => 'integer',
+        ]);
         $role->update($request->all());
         $role->permissions()->sync($request->input('permissions', []));
 
@@ -82,8 +90,14 @@ class RolesController extends Controller
         return back();
     }
 
-    public function massDestroy(MassDestroyRoleRequest $request)
+    public function massDestroy(Request $request)
     {
+        abort_unless(\Gate::allows('role_delete'), 403);
+
+        $request->validate([
+            'ids' => 'required|array',
+            'ids.*' => 'exists:roles,id',
+        ]);
         Role::whereIn('id', request('ids'))->delete();
 
         return response(null, 204);
